chore: move shared module to shared folder
