+++ src/app/auth/auth.module.ts
@@ -4,7 +4,7 @@ import { NgxsModule } from "@ngxs/store";
 
 import { PipesModule } from "@/pipes/pipes.module";
 
-import { SharedModule } from "../shared.module";
+import { SharedModule } from "../shared/shared.module";
 import { AuthLockedComponent } from "./auth-locked/auth-locked.component";
 import { AuthPinComponent } from "./auth-pin/auth-pin.component";
 import { AuthTouchIdComponent } from "./auth-touch-id/auth-touch-id.component";
