+++ src/app/components/input-address/input-address.stories.ts
@@ -8,7 +8,7 @@ import {
 import { IonicModule } from "@ionic/angular";
 import { moduleMetadata, storiesOf } from "@storybook/angular";
 
-import { SharedModule } from "@/app/shared.module";
+import { SharedModule } from "@/app/shared/shared.module";
 import { DirectivesModule } from "@/directives/directives.module";
 import { TruncateMiddlePipe } from "@/pipes/truncate-middle/truncate-middle";
