+++ src/app/components/viewer-log/viewer-log.component.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { IonicModule } from "@ionic/angular";
 
-import { SharedModule } from "@/app/shared.module";
+import { SharedModule } from "@/app/shared/shared.module";
 
 import { ViewerLogComponent } from "./viewer-log.component";
