+++ src/app/components/wallet-picker/wallet-picker.modal.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { IonicModule } from "@ionic/angular";
 
-import { SharedModule } from "@/app/shared.module";
+import { SharedModule } from "@/app/shared/shared.module";
 
 import { WalletPickerComponentModule } from "./wallet-picker.component.module";
 import { WalletPickerModal } from "./wallet-picker.modal";
