+++ src/app/components/wallet-picker/wallet-picker.stories.ts
@@ -1,7 +1,7 @@
 import { IonicModule } from "@ionic/angular";
 import { moduleMetadata, storiesOf } from "@storybook/angular";
 
-import { SharedModule } from "@/app/shared.module";
+import { SharedModule } from "@/app/shared/shared.module";
 import { Contact } from "@/models/model";
 import { PipesModule } from "@/pipes/pipes.module";
