--- src/app/shared.module.ts
@@ -1,9 +1,0 @@
-import { CommonModule } from "@angular/common";
-import { NgModule } from "@angular/core";
-import { FormsModule, ReactiveFormsModule } from "@angular/forms";
-import { TranslateModule } from "@ngx-translate/core";
-
-@NgModule({
-	exports: [CommonModule, FormsModule, ReactiveFormsModule, TranslateModule],
-})
-export class SharedModule {}
